pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pricing-rl"
version = "0.1.0"
description = "RL-based dynamic pricing agent for foodservice distribution"
requires-python = ">=3.11"
dependencies = [
    "stable-baselines3>=2.3.0",
    "sb3-contrib>=2.3.0",
    "gymnasium>=0.29.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "streamlit>=1.30.0",
    "plotly>=5.18.0",
    "tensorboard>=2.15.0",
    "torch>=2.1.0",
    "scipy>=1.11.0",
    "anthropic>=0.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
